+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 - raid content type is now only disabled for Party frames
 - the sound selection under Settings is now scrollable, containing 20 items before having to scroll
 - synchronize min gap with max frame dimensions
-- added new setting to show fractions of cast durations
+- added a new setting to show fractions of cast durations
 - added support to use DandersFrames if present
+- added a new setting that briefly highlights interrupted spells - only works for non-channels
 
 ## Version 1.0.0-alpha4